+++ extension-files/bringweb3-sdk/bringInitBackground.ts
@@ -1,5 +1,4 @@
 import { openExtensionCashbackPage } from './utils/background/openExtensionCashbackPage';
-import fetchDomains from "./utils/api/fetchDomains.js"
 import validateDomain from "./utils/api/validateDomain.js"
 import { ApiEndpoint } from "./utils/apiEndpoint.js"
 import parseUrl from "./utils/parseUrl.js"
@@ -12,33 +11,11 @@ import getWalletAddress from "./utils/background/getWalletAddress.js"
 import sendMessage from "./utils/background/sendMessage.js"
 import getUserId from "./utils/background/getUserId.js"
 import showNotification from "./utils/background/showNotification.js"
+import isWhitelisted from './utils/background/isWhitelisted';
+import { updateCache } from './utils/background/updateCache';
 
-const calcDelay = (timestamp: number) => {
-    const now = Date.now()
-    return (timestamp - now) / 1000 / 60 // milliseconds to minutes
-}
-
-const updateCache = async (apiKey: string) => {
-    const res = await fetchDomains(apiKey)
-
-    storage.set('relevantDomains', res.relevantDomains)
-
-    const { nextUpdateTimestamp } = res
-
-    const delay = calcDelay(nextUpdateTimestamp)
-
-    chrome.alarms.create(UPDATE_CACHE_ALARM_NAME, {
-        delayInMinutes: delay || 60 * 24 * 2
-    })
-    return res.relevantDomains
-}
-
-const getRelevantDomain = async (url: string | undefined, apiKey: string) => {
-    let relevantDomains = await storage.get('relevantDomains')
-
-    if (relevantDomains === undefined) {
-        relevantDomains = await updateCache(apiKey)
-    }
+const getRelevantDomain = async (url: string | undefined) => {
+    const relevantDomains = await updateCache()
 
     if (!url || !relevantDomains || !relevantDomains.length) return ''
     const domain = parseUrl(url)
@@ -65,6 +42,7 @@ const urlsDict: UrlDict = {}
 interface Configuration {
     identifier: string
     apiEndpoint: string
+    whitelistEndpoint?: string
     cashbackPagePath?: string
 }
 /**
@@ -75,6 +53,7 @@ interface Configuration {
  * @param {Object} configuration - The configuration object.
  * @param {string} configuration.identifier - The identifier for the extension.
  * @param {string} configuration.apiEndpoint - The API endpoint ('prod' or 'sandbox').
+ * @param {string} configuration.whitelistEndpoint - Endpoint for whitelist of redirect urls.
  * @param {string} [configuration.cashbackPagePath] - Optional path to the cashback page.
  * @throws {Error} Throws an error if identifier or apiEndpoint is missing, or if apiEndpoint is invalid.
  * @returns {Promise<void>}
@@ -97,25 +76,26 @@ interface Configuration {
  * bringInitBackground({
  *   identifier: '<bring_identifier>',
  *   apiEndpoint: 'sandbox',
+ *   whitelistEndpoint: 'https://example.com/whitelist.json',
  *   cashbackPagePath: '/cashback.html'
  * });
  */
-const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }: Configuration) => {
+const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath, whitelistEndpoint }: Configuration) => {
     if (!identifier || !apiEndpoint) throw new Error('Missing configuration')
     if (!['prod', 'sandbox'].includes(apiEndpoint)) throw new Error('unknown apiEndpoint')
 
     ApiEndpoint.getInstance().setApiEndpoint(apiEndpoint)
+    ApiEndpoint.getInstance().setWhitelistEndpoint(whitelistEndpoint || '')
+    ApiEndpoint.getInstance().setApiKey(identifier)
 
-    if (!await storage.get('relevantDomains')) {
-        updateCache(identifier)
-    }
+    updateCache()
 
     chrome.alarms.onAlarm.addListener(async (alarm) => {
         const { name } = alarm
 
         switch (name) {
             case UPDATE_CACHE_ALARM_NAME:
-                updateCache(identifier)
+                updateCache()
                 break;
             default:
                 console.error('alarm with no use case:', name);
@@ -131,8 +111,8 @@ const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }
 
         switch (action) {
             case 'ACTIVATE': {
-                const { domain, extensionId, time } = request
-                handleActivate(domain, extensionId, identifier, cashbackPagePath, time)
+                const { domain, extensionId, time, redirectUrl } = request
+                handleActivate(domain, extensionId, identifier, cashbackPagePath, time, sender.tab?.id, redirectUrl)
                     .then(() => sendResponse());
                 return true;
             }
@@ -172,7 +152,8 @@ const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }
                 return true;
             }
             case 'OPEN_CASHBACK_PAGE':
-                openExtensionCashbackPage(cashbackPagePath || '')
+                const { url } = request
+                openExtensionCashbackPage(url || cashbackPagePath)
                 sendResponse({ message: 'cashback page opened successfully' })
                 return true
         }
@@ -199,7 +180,7 @@ const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }
 
         urlsDict[tabId] = url
 
-        const match = await getRelevantDomain(tab.url, identifier);
+        const match = await getRelevantDomain(tab.url);
 
         if (!match || !match.length) {
             await showNotification(identifier, tabId, cashbackPagePath, url)
@@ -208,7 +189,7 @@ const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }
 
         const address = await getWalletAddress(tabId);
 
-        const { token, isValid, iframeUrl } = await validateDomain({
+        const { token, isValid, iframeUrl, networkUrl } = await validateDomain({
             apiKey: identifier,
             body: {
                 domain: match,
@@ -221,6 +202,8 @@ const bringInitBackground = async ({ identifier, apiEndpoint, cashbackPagePath }
             if (isValid === false) addQuietDomain(match);
             return;
         }
+        if (!await isWhitelisted(networkUrl, await storage.get('redirectsWhitelist'))) return;
+
         const userId = await getUserId()
         sendMessage(tabId, {
             action: 'INJECT',