+++ extension-files/bringweb3-sdk/utils/api/fetchDomains.ts
@@ -1,8 +1,20 @@
+import { ApiEndpoint } from "../apiEndpoint"
 import apiRequest from "./apiRequest"
 
 const fetchDomains = async (apiKey: string) => {
+    const whitelistEndpoint = ApiEndpoint.getInstance().getWhitelistEndpoint()
 
-    const res = await apiRequest({ path: '/domains', apiKey, method: 'GET' })
+    const request: Parameters<typeof apiRequest>[0] = {
+        path: '/domains',
+        apiKey,
+        method: 'GET',
+    }
+
+    if (whitelistEndpoint) {
+        request.params = { whitelist: encodeURIComponent(whitelistEndpoint) }
+    }
+
+    const res = await apiRequest(request)
 
     return res
 }