+++ extension-files/bringweb3-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @bringweb3/chrome-extension-kit
 
+## 1.3.3
+
+### Patch Changes
+
+- 6ac9ee2: fix whitelist check
+
+## 1.3.2
+
+### Patch Changes
+
+- a151374: Fix no whitelist usecase
+
+## 1.3.1
+
+### Patch Changes
+
+- 2b63c79: fix relevant domains list not updating
+
+## 1.3.0
+
+### Minor Changes
+
+- b406236: Add option for whitelisted redirect urls
+
+### Patch Changes
+
+- ea6057f: fix notification link
+- ed6dcce: Disable http caching for whitelist call
+
 ## 1.2.14
 
 ### Patch Changes